1.别踩白格多人玩法开发

+++ Assets/Develop/FGUFW/Core/Layer2/System/NetworkSyncSystem/INetworkSyncSystem.cs
@@ -5,7 +5,7 @@ namespace FGUFW.Core
 {
     public interface INetworkSyncSystem : ISystem
     {
-        void SendMsg(uint cmd,ushort gameplayID,object msg);
+        void SendMsg(uint cmd,long gameplayID,object msg);
         IMessenger<uint,ByteString> Messenger{get;}
     }
 }
